Renamed ModynModels to CoresetMethodsSupport
